Backend/requirements.txt
flask
torch
torchvision
pillow
gdown
matplotlib